+++ lib/spear/writing.ex
@@ -12,14 +12,20 @@ defmodule Spear.Writing do
       delete_req: 1,
       delete_req_options: 1,
       tombstone_req: 1,
-      tombstone_req_options: 1
+      tombstone_req_options: 1,
+      batch_append_req: 1,
+      batch_append_req_options: 1,
+      batch_append_req_proposed_message: 0
     ]
 
   import Spear.Records.Shared,
     only: [
       stream_identifier: 1,
-      empty: 0
+      empty: 0,
+      uuid: 1
     ]
+
+  require Spear.Records.Google, as: Google
 
   alias Spear.ExpectationViolation
 
@@ -51,6 +57,15 @@ defmodule Spear.Writing do
           stream_identifier: stream_identifier(stream_name: params.stream),
           expected_stream_revision: map_expectation(params.expect)
         )
+    )
+  end
+
+  def build_batch_append_request(params) do
+    batch_append_req(
+      correlation_id: uuid(value: {:string, params.batch_id}),
+      options: map_batch_append_options(params),
+      proposed_messages: Enum.map(params.events, &map_batch_append_messages/1),
+      is_final: params.done?
     )
   end
 
@@ -102,6 +117,43 @@ defmodule Spear.Writing do
       }
       |> json_encode!.(),
       content_type: "application/vnd.eventstore.events+json"
+    )
+  end
+
+  defp map_batch_append_options(%{continuation?: true}), do: :undefined
+
+  defp map_batch_append_options(params) do
+    batch_append_req_options(
+      stream_identifier: stream_identifier(stream_name: params.stream_name),
+      expected_stream_position: map_expected_position(params.expect),
+      deadline: map_deadline(params.deadline)
+    )
+  end
+
+  defp map_deadline(nil), do: :undefined
+  # coveralls-ignore-start
+  defp map_deadline(deadline), do: Google.timestamp(deadline)
+  # coveralls-ignore-stop
+
+  defp map_expected_position(revision) when is_integer(revision) and revision >= 1,
+    do: {:stream_position, revision}
+
+  defp map_expected_position(:empty), do: {:no_stream, Google.empty()}
+  defp map_expected_position(:exists), do: {:stream_exists, Google.empty()}
+  defp map_expected_position(:any), do: {:any, Google.empty()}
+
+  # coveralls-ignore-start
+  defp map_batch_append_messages(batch_append_req_proposed_message() = message) do
+    message
+  end
+
+  # coveralls-ignore-stop
+
+  defp map_batch_append_messages(%Spear.Event{} = event) do
+    Spear.Event.to_proposed_message(
+      event,
+      %{"application/json" => &Jason.encode!/1},
+      :batch_append
     )
   end
 end